pyproject.toml
[project]
name = "python-ai-samples"
version = "0.1.0"
description = "Sample apps for building AI projects with Python and hosting them on Azure"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "azure-ai-inference>=1.0.0b9",
    "azure-identity>=1.24.0",
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
